refactoring / NoteGraph
